Split inspect up into individual source files for each command.
Also, break utility stuff that the commands will need out into their own
files and move inspect.hs into its own source directory.  This also
means I can hide those option data types.
